+++ src/main/java/iKeeper/iKeeper/Homepage/model/entity/Field.java
@@ -1,4 +1,4 @@
-package iKeeper.iKeeper.Homepage.model;
+package iKeeper.iKeeper.Homepage.model.entity;
 
 import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
