+++ src/main/java/iKeeper/iKeeper/Homepage/repository/CalendarRepository.java
@@ -1,6 +1,6 @@
 package iKeeper.iKeeper.Homepage.repository;
 
-import iKeeper.iKeeper.Homepage.model.Calendar;
+import iKeeper.iKeeper.Homepage.model.entity.Calendar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CalendarRepository extends JpaRepository<Calendar, Long> {
