+++ src/main/java/iKeeper/iKeeper/Homepage/repository/FieldRepository.java
@@ -1,6 +1,6 @@
 package iKeeper.iKeeper.Homepage.repository;
 
-import iKeeper.iKeeper.Homepage.model.Field;
+import iKeeper.iKeeper.Homepage.model.entity.Field;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FieldRepository extends JpaRepository<Field, Long> {
