+++ src/test/java/iKeeper/iKeeper/Homepage/service/CalendarServiceTest.java
@@ -1,13 +1,8 @@
 package iKeeper.iKeeper.Homepage.service;
 
-import iKeeper.iKeeper.Homepage.model.Calendar;
-import iKeeper.iKeeper.Homepage.model.request.CalendarCreationRequest;
-import iKeeper.iKeeper.Homepage.repository.CalendarRepository;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ExtendWith(SpringExtension.class)
